python-ai/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
google-genai>=0.8.0
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
